refactor (featured-program): move modal body to program-detail component
