+++ components/AboutJabar/FeaturedProgram/index.vue
@@ -4,17 +4,9 @@
       <div class="py-8 px-10 rounded-xl bg-white">
         <div class="p-4">
           <!-- Search and Filter -->
-          <section class="flex w-full justify-between mb-8">
+          <section class="flex w-full justify-end mb-8">
             <!--
-              NOTE:
-              Category Selection will be implemented in the next PR
-
-            <div class="inline-flex gap-4 items-center">
-              <p class="font-lato text-md text-gray-800 font-normal">
-                Tampilkan
-              </p>
-              <jds-select />
-            </div>
+              TODO: Add category selection here
             -->
 
             <!--
@@ -40,126 +32,10 @@
           </section>
 
           <Modal :show="openModal">
-            <div class="grid grid-cols-2 w-full px-6 gap-y-5">
-              <section class="col-span-2">
-                <figure class="inline-flex w-10 h-10 p-1 items-center justify-center mb-1">
-                  <img
-                    src="/icons/program-unggulan/sapawarga.svg"
-                    :alt="programDetail.title"
-                    class="w-full h-full object-center"
-                  >
-                </figure>
-                <h1 class="font-medium text-2xl text-green-700 leading-relaxed">
-                  {{ programDetail.title || '-' }}
-                </h1>
-              </section>
-
-              <section class="col-span-2 flex gap-4">
-                <Icon src="/icons/info-circle-outline.svg" size="16px" class="self-start text-green-600" />
-                <div>
-                  <h2 class="font-lato text-xs text-blue-gray-200 mb-1 leading-5">
-                    Deskripsi Program
-                  </h2>
-                  <p class="text-gray-800 font-normal text-sm leading-relaxed">
-                    {{ programDetail.description || '-' }}
-                  </p>
-                </div>
-              </section>
-
-              <section class="col-span-1 flex gap-4">
-                <Icon src="/icons/organization.svg" size="16px" class="self-start text-green-600" />
-                <div>
-                  <h2 class="font-lato text-xs text-blue-gray-200 mb-1 leading-5">
-                    Organisasi
-                  </h2>
-                  <p class="text-gray-800 font-normal text-sm leading-relaxed">
-                    {{ programDetail.organization || "-" }}
-                  </p>
-                </div>
-              </section>
-
-              <section class="col-span-1 flex gap-4">
-                <Icon src="/icons/category.svg" size="16px" class="self-start text-green-600" />
-                <div>
-                  <h2 class="font-lato text-xs text-blue-gray-200 mb-1 leading-5">
-                    Kategori
-                  </h2>
-                  <template v-if="hasCategory">
-                    <p
-                      v-for="(category, index) in programDetail.categories"
-                      :key="index"
-                      class="text-gray-800 font-normal capitalize text-sm leading-relaxed"
-                    >
-                      {{ category }}
-                    </p>
-                  </template>
-                  <template v-else>
-                    <p class="text-gray-800 font-normal text-sm leading-relaxed">
-                      -
-                    </p>
-                  </template>
-                </div>
-              </section>
-
-              <section class="col-span-2 flex gap-4">
-                <Icon src="/icons/link.svg" size="16px" class="self-start text-green-600" />
-                <div>
-                  <h2 class="font-lato text-xs text-blue-gray-200 mb-1 leading-5">
-                    Link Website
-                  </h2>
-                  <template v-if="hasWebsite">
-                    <a
-                      v-for="(website, index) in programDetail.websites"
-                      :key="index"
-                      :href="website"
-                      rel="noopener noreferrer"
-                      target="_blank"
-                      class="block text-blue-500 text-sm hover:text-blue-700 underline mb-1 leading-relaxed"
-                    >
-                      {{ website }}
-                    </a>
-                  </template>
-                  <template v-else>
-                    <p class="text-gray-800 font-normal text-sm leading-relaxed">
-                      Link tidak tersedia
-                    </p>
-                  </template>
-                </div>
-              </section>
-
-              <section v-if="hasSocialSites" class="col-span-2 flex gap-4">
-                <Icon src="/icons/share.svg" size="16px" class="self-start text-green-600" />
-                <div>
-                  <h2 class="font-lato text-xs text-blue-gray-200 mb-1 leading-5">
-                    Social Media
-                  </h2>
-                  <div class="flex gap-2">
-                    <Link
-                      v-for="(site, index) in socialSites"
-                      :key="index"
-                      :link="site.link"
-                      class="flex flex-col p-2 items-center justify-center gap-1
-                      cursor-pointer hover:bg-green-50"
-                    >
-                      <Icon
-                        :src="site.logo"
-                        height="18px"
-                      />
-                      <p class="text-gray-800 capitalize font-normal text-sm leading-relaxed">
-                        {{ site.label }}
-                      </p>
-                    </Link>
-                  </div>
-                </div>
-              </section>
-
-              <section
-                class="col-span-2 flex justify-center mb-4"
-                @click="closeDetail"
-              >
-                <Button>Tutup</Button>
-              </section>
-            </div>
+            <ProgramDetail
+              :program-detail="programDetail"
+              @close="closeDetail"
+            />
           </Modal>
         </div>
       </div>
@@ -168,12 +44,14 @@
 </template>
 
 <script>
+import ProgramDetail from './ProgramDetail'
 import Card from '~/components/Base/Card'
 
 export default {
   name: 'FeaturedProgramList',
   components: {
-    Card
+    Card,
+    ProgramDetail
   },
   data () {
     return {
@@ -190,85 +68,6 @@ export default {
     } catch (error) {
       this.programList = []
       // silent error
-    }
-  },
-  computed: {
-
-    /**
-     * Check whether selected program has website
-     */
-    hasWebsite () {
-      const { websites } = this.programDetail
-
-      if (Array.isArray(websites) && websites.length) {
-        return true
-      }
-
-      return false
-    },
-
-    /**
-     * Check whether selected program has category
-     * @returns {Boolean}
-     */
-    hasCategory () {
-      const { categories } = this.programDetail
-
-      if (Array.isArray(categories) && categories.length) {
-        return true
-      }
-
-      return false
-    },
-
-    /**
-     * Check whether selected program has social sites
-     * @returns {Boolean}
-     */
-    hasSocialSites () {
-      // eslint-disable-next-line camelcase
-      const { social_media } = this.programDetail
-
-      // eslint-disable-next-line camelcase
-      if (social_media && Object.keys(social_media).length) {
-        const socialMedia = Object.values(social_media)
-
-        if (socialMedia.every(item => item === '')) {
-          return false
-        }
-
-        return true
-      }
-
-      return false
-    },
-
-    /**
-     * Array of available social media sites
-     * based on selected program
-     * @returns {Array}
-     */
-    socialSites () {
-      if (this.hasSocialSites) {
-        // eslint-disable-next-line camelcase
-        const { social_media } = this.programDetail
-        const socialMedia = []
-
-        // eslint-disable-next-line camelcase
-        for (const item in social_media) {
-          if (Object.hasOwnProperty.call(social_media, item)) {
-            socialMedia.push({
-              logo: `/icons/social-media/${item}-logo.svg`,
-              label: item,
-              link: social_media[item]
-            })
-          }
-        }
-
-        return socialMedia.filter(item => item.link !== '')
-      }
-
-      return []
     }
   },
   methods: {
